Created new test class.
This was easier to control than the bash (escaping values in a variable was
a nightmare..!).  It also means I can have a 'exp,val' where value is true
if the expression should pass meaning it is self testing.  At the end, it
will display all the expressions that failed.