Task #2 - Literal list is working.
There is a wasted state for the first literal in the list, but it
makes the code easier for the rest of the list.  I've only tested
this with [a], [ab] and [abc] and it seems to be working.  I need
to test the rest of the expressions together to make sure there
are no bugs.

For now, I've changed it to make the correct output for the search